# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Here we handle Geolocation!",
        "title": "Vio Challenge Server",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/geolocations/{ip_address}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geolocation"
                ],
                "summary": "Get geolocation by IP address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ip address",
                        "name": "ip_address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Geolocation"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.Geolocation": {
            "type": "object",
            "required": [
                "city",
                "country",
                "country_code",
                "ip_address",
                "latitude",
                "longitude",
                "mystery_value"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "maxLength": 255
                },
                "country": {
                    "type": "string",
                    "maxLength": 255
                },
                "country_code": {
                    "type": "string"
                },
                "ip_address": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "mystery_value": {
                    "type": "string"
                }
            }
        }
    }
}